Youth application controller done.
