--- .coverage
@@ -1,1 +1,0 @@
-!coverage.py: This is a private format, don't read it directly!{"lines": {"/home/ben/paperless_registration/models/AdultApplications.py": [], "/home/ben/paperless_registration/models/CharterApplications.py": [], "/home/ben/paperless_registration/models/sample_data.py": [], "/home/ben/paperless_registration/models/tests.py": [], "/home/ben/paperless_registration/models/User.py": [], "/home/ben/paperless_registration/models/YouthApplications.py": [], "/home/ben/paperless_registration/models/Youth.py": [], "/home/ben/paperless_registration/models/__init__.py": [], "/home/ben/paperless_registration/models/Base.py": [], "/home/ben/paperless_registration/models/create_sample_data.py": [], "/home/ben/paperless_registration/models/Districts.py": [], "/home/ben/paperless_registration/models/Subdistricts.py": [], "/home/ben/paperless_registration/models/clear_sample_data.py": [], "/home/ben/paperless_registration/models/Units.py": [], "/home/ben/paperless_registration/models/SponsoringOrganizations.py": [], "/home/ben/paperless_registration/models/Volunteers.py": [], "/home/ben/paperless_registration/models/Guardians.py": []}}
