--- controllers/AdultApplication.py
@@ -1,1 +1,0 @@
-# TODO: implement this
