+++ controllers/__init__.py
@@ -2,7 +2,8 @@
 
 Controllers are read-only; they can read from the DB, but never write to it.
 """
-# TODO: add license
-# TODO: add Google-riffic docs everywhere
-# TODO: code coverage
-# TODO: add data integrity check scripts
+
+
+class InvalidActionException(Exception):
+    """ The action attempted is not valid """
+    pass
