+++ models/YouthApplications.py
@@ -33,33 +33,6 @@ class YouthApplications(Base.Object):  # pylint: disable=too-many-instance-attri
         self.youth_id = ''
         self.scoutnet_id = 0
 
-    def set_status(self, status):
-        """ Set the application status """
-        self.status = status
-
-    def set_guardian_approval(self, approval):
-        """ Set guardian approval """
-        self.guardian_approval_guardian_id = approval['guardian_approval_guardian_id']
-        self.guardian_approval_signature = approval['guardian_approval_signature']
-        self.guardian_approval_date = approval['guardian_approval_date']
-
-    def set_unit_approval(self, approval):
-        """ Set unit approval """
-        self.unit_approval_user_id = approval['unit_approval_user_id']
-        self.unit_approval_signature = approval['unit_approval_signature']
-        self.unit_approval_date = approval['unit_approval_date']
-
-    def set_fee_payment(self, data):
-        """ Set fee payment """
-        self.fee_payment_date = data['fee_payment_date']
-        self.fee_payment_user_id = data['fee_payment_user_id']
-        self.fee_payment_receipt = data['fee_payment_receipt']
-
-    def set_recorded_in_scoutnet(self, data):
-        """ Mark as recorded in ScoutNet """
-        self.scoutnet_id = data['scoutnet_id']
-        self.recorded_in_scoutnet_date = data['date']
-
     @staticmethod
     def get_uuid_prefix():
         return 'yap'
@@ -125,12 +98,7 @@ class StatusValidator(object):  # pylint: disable=too-few-public-methods
                 'unit_approval_date': Base.FIELD_REQUIRED,
             })
         if self.status in [STATUS_READY_FOR_SCOUTNET, STATUS_COMPLETE]:
-            fields.update({
-                # TODO: not required for LDS units; make conditional?
-                # 'fee_payment_date': Base.FIELD_REQUIRED,
-                # 'fee_payment_user_id': Base.FIELD_REQUIRED,
-                # 'fee_payment_receipt': Base.FIELD_REQUIRED,
-            })
+            fields.update({})
         if self.status in [STATUS_COMPLETE]:
             fields.update({
                 'recorded_in_scoutnet_date': Base.FIELD_REQUIRED,
