+++ tools/clear_sample_data.py
