+++ tools/sample_data.py
